The byte value for traffic is now a float because announced sometimes decides that 0.000002 bytes should be possible

+++ src/github.com/dereulenspiegel/node-informant/gluon-collector/data/statistics.go
@@ -13,9 +13,9 @@ type MeshVPNStruct struct {
 }
 
 type TrafficObject struct {
-	Bytes   uint64 `json:"bytes,omitempty"`
-	Packets uint64 `json:"packets,omitempty"`
-	Dropped uint64 `json:"dropped,omitempty"`
+	Bytes   float64 `json:"bytes,omitempty"`
+	Packets uint64  `json:"packets,omitempty"`
+	Dropped uint64  `json:"dropped,omitempty"`
 }
 
 type TrafficStruct struct {
